extract ShellTerminal to own component
