+++ ui/js/components/index.js
@@ -2,3 +2,4 @@ export CounterButton from './CounterButton/CounterButton';
 export WebSocketStatus from './WebSocketStatus/WebSocketStatus';
 export HostDisplay from './Hosts/HostDisplay';
 export ConfigurationDisplay from './ConfigurationDisplay/ConfigurationDisplay';
+export ShellTerminal from './ShellTerminal/ShellTerminal';
